# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Esta es la API de RPL 2.0 para todos y todas",
        "version": "0.1",
        "title": "RPL 2.0",
        "termsOfService": "Terms of service",
        "contact": {
            "name": "Alejandro Levinas o Mat\u00edas Cano",
            "url": "http://example.com/support",
            "email": "levinasale@gmail.com;matiasjosecc@gmail.com"
        },
        "license": {
            "name": "License of API",
            "url": "http://example.com/support"
        }
    },
  "host": "localhost:38497",
    "basePath": "/",
...
    "paths": {
...
        "/api/auth/profile": {
            "get": {
                "tags": [
                    "authentication-controller"
                ],
                "summary": "getUser",
                "operationId": "getUserUsingGET",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "accountNonExpired",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "accountNonLocked",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "authorities[0].authority",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "credentialsNonExpired",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "enabled",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "required": false,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.dateCreated",
                        "in": "query",
                        "required": false,
                        "type": "string",
                        "format": "date-time"
                    },
                    {
                        "name": "user.degree",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.email",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.emailValidated",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "user.id",
                        "in": "query",
                        "required": false,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "user.lastUpdated",
                        "in": "query",
                        "required": false,
                        "type": "string",
                        "format": "date-time"
                    },
                    {
                        "name": "user.name",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.password",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.studentId",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.surname",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.university",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.username",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "username",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/UserResponseDTO"
                        }
                    },
                    "403": {
                        "description": "Forbidden!!!!!"
                    },
                    "500": {
                        "description": "500 message",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                },
                "deprecated": false
            }
        },
...
        "/api/courses": {
            "get": {
                "tags": [
                    "courses-controller"
                ],
                "summary": "getCourses",
                "operationId": "getCoursesUsingGET",
                "produces": [
                    "*/*"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/CourseResponseDTO"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden!!!!!"
                    },
                    "500": {
                        "description": "500 message",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                },
                "deprecated": false
            },
            "post": {
                "tags": [
                    "courses-controller"
                ],
                "summary": "createCourse",
                "operationId": "createCourseUsingPOST",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "accountNonExpired",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "accountNonLocked",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "authorities[0].authority",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "createCourseRequestDTO",
                        "description": "createCourseRequestDTO",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateCourseRequestDTO"
                        }
                    },
                    {
                        "name": "credentialsNonExpired",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "enabled",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "required": false,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.dateCreated",
                        "in": "query",
                        "required": false,
                        "type": "string",
                        "format": "date-time"
                    },
                    {
                        "name": "user.degree",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.email",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.emailValidated",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "user.id",
                        "in": "query",
                        "required": false,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "user.lastUpdated",
                        "in": "query",
                        "required": false,
                        "type": "string",
                        "format": "date-time"
                    },
                    {
                        "name": "user.name",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.password",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.studentId",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.surname",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.university",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.username",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "username",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/CourseResponseDTO"
                        }
                    }
                },
                "deprecated": false
            }
        },
        "/api/courses/{courseId}": {
            "get": {
                "tags": [
                    "courses-controller"
                ],
                "summary": "getCourseDetails",
                "operationId": "getCourseDetailsUsingGET",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "accountNonExpired",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "accountNonLocked",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "authorities[0].authority",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "courseId",
                        "in": "path",
                        "description": "courseId",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "credentialsNonExpired",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "enabled",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "required": false,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.dateCreated",
                        "in": "query",
                        "required": false,
                        "type": "string",
                        "format": "date-time"
                    },
                    {
                        "name": "user.degree",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.email",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.emailValidated",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "user.id",
                        "in": "query",
                        "required": false,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "user.lastUpdated",
                        "in": "query",
                        "required": false,
                        "type": "string",
                        "format": "date-time"
                    },
                    {
                        "name": "user.name",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.password",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.studentId",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.surname",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.university",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user.username",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "username",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/UserPrincipal"
                        }
                    },
                    "403": {
                        "description": "Forbidden!!!!!"
                    },
                    "500": {
                        "description": "500 message",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                },
                "deprecated": false
            }
        },
...
    },
    "definitions": {
...
        "CourseResponseDTO": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "date_created": {
                    "type": "string",
                    "format": "date-time"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "img_uri": {
                    "type": "string"
                },
                "last_updated": {
                    "type": "string",
                    "format": "date-time"
                },
                "name": {
                    "type": "string"
                },
                "semester": {
                    "type": "string"
                },
                "university_course_id": {
                    "type": "string"
                }
            },
            "title": "CourseResponseDTO"
        },
...
        "CreateCourseRequestDTO": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "semester": {
                    "type": "string"
                },
                "university": {
                    "type": "string"
                },
                "university_course_id": {
                    "type": "string"
                }
            },
            "title": "CreateCourseRequestDTO"
        },
...
        "GrantedAuthority": {
            "type": "object",
            "properties": {
                "authority": {
                    "type": "string"
                }
            },
            "title": "GrantedAuthority"
        },
...
        "UserPrincipal": {
            "type": "object",
            "properties": {
                "account_non_expired": {
                    "type": "boolean"
                },
                "account_non_locked": {
                    "type": "boolean"
                },
                "authorities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/GrantedAuthority"
                    }
                },
                "credentials_non_expired": {
                    "type": "boolean"
                },
                "enabled": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "name": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            },
            "title": "UserPrincipal"
        },
        "UserResponseDTO": {
            "type": "object",
            "properties": {
                "date_created": {
                    "type": "string",
                    "format": "date-time"
                },
                "degree": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "email_validated": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "last_updated": {
                    "type": "string",
                    "format": "date-time"
                },
                "name": {
                    "type": "string"
                },
                "student_id": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                },
                "university": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            },
            "title": "UserResponseDTO"
        }
    }
}